// src/api/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        BranchModel: {
            /** Format: int32 */
            id?: number;
            name?: string;
            repositoryId?: string;
            /** Format: int32 */
            headChangeset?: number;
            /** Format: date-time */
            date?: string;
            owner?: string;
            comment?: string | null;
            type?: string;
            hasCodeReview?: boolean;
            /** Format: int32 */
            codeReviewsCount?: number;
            codeReviewIds?: number[] | null;
            /** Format: date-time */
            lastActivity?: string;
            headChangesetComment?: string | null;
            canBeDeleted?: boolean;
            canBeDeletedIncludingChangesets?: boolean;
            /** Format: int32 */
            changesetsCount?: number;
            /** Format: int32 */
            status?: number | null;
            isMain?: boolean;
            mergeInfo?: components["schemas"]["MergeInfoModel"][] | null;
            attributeValues?: components["schemas"]["AttributeValueModel"][] | null;
            isHidden?: boolean;
        };
// ...
        ChangesetModel: {
            /** Format: int32 */
            id?: number;
            /** Format: int32 */
            changesetId?: number;
            repositoryId?: string;
            /** Format: int32 */
            parentChangesetId?: number;
            branch?: string;
            /** Format: int64 */
            branchId?: number;
            /** Format: date-time */
            date?: string;
            /** Format: uuid */
            guid?: string;
            owner?: string;
            comment?: string | null;
            changesetType?: string;
            hasCodeReview?: boolean;
            /** Format: int32 */
            codeReviewsCount?: number;
            codeReviewIds?: number[] | null;
            labels?: string[] | null;
            canBeDeleted?: boolean | null;
            mergeInfo?: components["schemas"]["MergeInfoModel"][] | null;
            attributeValues?: components["schemas"]["AttributeValueModel"][] | null;
        };
        /** @enum {string} */
        ChangesetOrderByType: "id" | "creationDate";
        CheckInRequest: {
            statusIgnoreCase?: boolean;
            items?: string[] | null;
            comment?: string | null;
        };
// ...
        CheckinResponseModel: {
            /** Format: int64 */
            changesetId?: number;
            branchName?: string;
            /** Format: uuid */
            guid?: string;
            /** Format: date-time */
            timestamp?: string;
            repository?: string;
        };
// ...
        CodeReviewModel: {
            /** Format: int64 */
            id?: number;
            object?: components["schemas"]["CodeReviewObject"];
            assignee?: string | null;
            owner?: string;
            /** Format: date-time */
            timestamp?: string;
            /** Format: date-time */
            modifiedTimestamp?: string;
            status?: components["schemas"]["CodeReviewModel.StatusEnum"];
            title?: string;
            description?: string | null;
            descriptionComment?: components["schemas"]["CodeReviewCommentModel"];
            /** Format: int32 */
            commentsCount?: number;
            /** Format: int32 */
            issuesCount?: number | null;
            /** Format: int32 */
            unansweredQuestionsCount?: number | null;
            /** Format: int32 */
            userCommentsCount?: number | null;
            comments?: components["schemas"]["CodeReviewCommentModel"][] | null;
            reviewers?: components["schemas"]["CodeReviewerModel"][] | null;
            mergeAllowed?: boolean | null;
            mergeIssues?: string | null;
            mergeInfo?: components["schemas"]["MergeInfoModel"][] | null;
            repositoryName?: string | null;
        };
// ...
        DataType: "Dir" | "BinaryFile" | "Symlink" | "TextFile";
// ...
        DiffModel: {
            author?: string;
            comment?: string | null;
            /** Format: date-time */
            timestamp?: string;
            source?: string;
            destination?: string;
            /** Format: int64 */
            destinationId?: number;
            objectType?: components["schemas"]["DiffObject"];
            repository?: string;
            diffs?: components["schemas"]["DiffCategoryModel"][] | null;
        };
// ...
        HistoryModel: {
            name?: string;
            prevName?: string | null;
            /** Format: int64 */
            changeset?: number;
            owner?: string;
            /** Format: date-time */
            date?: string;
            branch?: string;
            /** Format: int64 */
            branchId?: number;
            comment?: string | null;
            /** Format: int64 */
            revision?: number;
            /** Format: int64 */
            parentRevision?: number;
            type?: components["schemas"]["HistoryType"];
            /** Format: int64 */
            size?: number;
        };
// ...
        LabelModel: {
            /** Format: int32 */
            id?: number;
            name?: string;
            repositoryId?: string;
            /** Format: int32 */
            changeset?: number;
            /** Format: date-time */
            date?: string;
            owner?: string;
            comment?: string | null;
            type?: string;
            /** Format: uuid */
            guid?: string;
            hasCodeReview?: boolean;
            /** Format: int32 */
            codeReviewsCount?: number;
            codeReviewIds?: number[] | null;
            attributeValues?: components["schemas"]["AttributeValueModel"][] | null;
        };
// ...
        LoginRequest: {
            /** @description Username */
            user: string;
            /** @description User password */
            password: string;
        };
        LoginResponse: {
            accessToken?: string;
            refreshToken?: string;
        };
// ...
        StatusChangeType: "none" | "private" | "added" | "deleted" | "locallyDeleted" | "changed" | "checkedOut" | "copied" | "replaced" | "moved" | "ignored";
        StatusResponse: {
            workspaceVersion?: components["schemas"]["WorkspaceVersionResponse"];
            changes?: components["schemas"]["StatusResponse.Change"][] | null;
        };
        "StatusResponse.Change": {
            changeType?: components["schemas"]["StatusChangeType"];
            path?: string | null;
            sourcePath?: string | null;
            dataType?: components["schemas"]["DataType"];
            newContentOnCloudWorkspace?: components["schemas"]["StatusResponse.Change.CloudWorkspaceContentInfo"];
            oldContentOnUVCS?: components["schemas"]["StatusResponse.Change.UvcsContentInfo"];
            /** Format: int32 */
            newFsProtection?: number | null;
            /** Format: int32 */
            oldFsProtection?: number | null;
        };
// ...
        TreeResponse: {
            tree?: components["schemas"]["TreeResponse.TreeEntry"];
        };
        "TreeResponse.TreeEntry": {
            statusType?: components["schemas"]["StatusType"];
            contentOnUVCS?: components["schemas"]["TreeResponse.TreeEntry.UvcsContentInfo"];
            name?: string | null;
            /** Format: uuid */
            revision?: string;
            /** Format: int64 */
            revisionId?: number;
            /** Format: int64 */
            size?: number;
            type?: components["schemas"]["DataType"];
            /** Format: date-time */
            utcTimeStamp?: string | null;
            children?: components["schemas"]["TreeResponse.TreeEntry"][] | null;
            owner?: string | null;
            /** Format: int64 */
            itemId?: number;
        };
// ...
        WorkspaceResponse: {
            name?: string | null;
            /** Format: uuid */
            guid?: string;
            owner?: string | null;
            id?: string | null;
            /** Format: date-time */
            lastActivityDate?: string;
            repositoryName?: string | null;
            uvcsConnections?: components["schemas"]["WorkspaceResponse.UvcsConnection"][] | null;
            shares?: components["schemas"]["WorkspaceResponse.Share"][] | null;
            /** Format: uuid */
            projectGuid?: string | null;
        };
// ...
        WorkspacesResponse: {
            /** Format: int32 */
            offset?: number | null;
            /** Format: int32 */
            limit?: number | null;
            /** Format: int32 */
            count?: number;
            workspaces?: components["schemas"]["WorkspaceResponse"][] | null;
        };
        WorkspaceVersionResponse: {
            /** Format: uuid */
            headChangeset?: string;
            /** Format: int64 */
            lastGrantedAccessModifiedTime?: number;
            uvcsConnections?: components["schemas"]["WorkspaceVersionResponse.UvcsConnection"][] | null;
        };
        "WorkspaceVersionResponse.UvcsConnection": {
            /** Format: uuid */
            id?: string;
            /** Format: int32 */
            version?: number;
            /** Format: uuid */
            headChangeset?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}